src/test/java/uk/gov/hmcts/reform/orgrolemapping/controller/WelcomeControllerTest.java
package uk.gov.hmcts.reform.orgrolemapping.controller;

import org.junit.jupiter.api.Test;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import uk.gov.hmcts.reform.orgrolemapping.controller.advice.ErrorConstants;

import static org.junit.jupiter.api.Assertions.assertEquals;

class WelcomeControllerTest {

    private final WelcomeController sut = new WelcomeController();

    @Test
    void welcome() {
        assertEquals("Welcome to Organisation Role Mapping Service", sut.welcome());
    }

    @Test
    void functionalSleepTest() {
        ResponseEntity<Object> response =
                ResponseEntity.status(HttpStatus.OK).body("Sleep time for Functional tests is over");

        assertEquals(response, sut.waitFor(null));
    }

    @Test
    void errorConstantTest() {
        assertEquals(202, ErrorConstants.ACCEPTED.getErrorCode());
        assertEquals("Accepted", ErrorConstants.ACCEPTED.getErrorMessage());
    }


}
